+++ tests/src/test/java/com/orientechnologies/orient/test/database/speed/LocalMTCreateDocumentSpeedTest.java
@@ -7,15 +7,18 @@ import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.metadata.security.OSecurity;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.serialization.serializer.record.binary.ORecordSerializerBinary;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Random;
-import java.util.concurrent.*;
+import java.util.concurrent.Callable;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
 
 /**
  * @author Andrey Lomakin (a.lomakin-at-orientdb.com)
@@ -23,17 +26,17 @@ import java.util.concurrent.*;
  */
 @Test
 public class LocalMTCreateDocumentSpeedTest {
-  private static final Random                   random          = new Random();
-  private ODatabaseDocumentTx                   database;
-  private Date                                  date            = new Date();
-  private CountDownLatch                        latch           = new CountDownLatch(1);
-  private List<Future>                          futures;
-  private volatile boolean                      stop            = false;
-  private ExecutorService                       executorService = Executors.newCachedThreadPool();
+  private static final Random random = new Random();
+  private ODatabaseDocumentTx database;
+  private Date date = new Date();
+  private CountDownLatch latch = new CountDownLatch(1);
+  private List<Future> futures;
+  private volatile boolean stop = false;
+  private ExecutorService executorService = Executors.newCachedThreadPool();
 
-  private final List<String>                    users           = new ArrayList<String>();
+  private final List<String> users = new ArrayList<String>();
 
-  private final OPartitionedDatabasePoolFactory poolFactory     = new OPartitionedDatabasePoolFactory();
+  private final OPartitionedDatabasePoolFactory poolFactory = new OPartitionedDatabasePoolFactory();
 
   @BeforeClass
   public void init() {
@@ -55,8 +58,7 @@ public class LocalMTCreateDocumentSpeedTest {
     }
 
     futures = new ArrayList<Future>();
-    for (int i = 0; i < 1; i++)
-      futures.add(executorService.submit(new Saver()));
+    for (int i = 0; i < 1; i++) futures.add(executorService.submit(new Saver()));
   }
 
   public void cycle() throws Exception {
@@ -67,8 +69,7 @@ public class LocalMTCreateDocumentSpeedTest {
 
     System.out.println("Stop insertion");
     long sum = 0;
-    for (Future<Long> future : futures)
-      sum += future.get();
+    for (Future<Long> future : futures) sum += future.get();
 
     System.out.println("Speed : " + (sum / futures.size()) + " ns per document.");
 
@@ -81,8 +82,11 @@ public class LocalMTCreateDocumentSpeedTest {
     System.out.println("Doc count : " + database.countClass("Account"));
 
     for (int i = 0; i < 8; i++)
-      futures.add(executorService.submit(new Reader(database.countClass("Account"), database.getMetadata().getSchema()
-          .getClass("Account").getDefaultClusterId())));
+      futures.add(
+          executorService.submit(
+              new Reader(
+                  database.countClass("Account"),
+                  database.getMetadata().getSchema().getClass("Account").getDefaultClusterId())));
 
     latch.countDown();
 
@@ -91,22 +95,19 @@ public class LocalMTCreateDocumentSpeedTest {
     stop = true;
 
     sum = 0;
-    for (Future future : futures)
-      sum += (Long) future.get();
+    for (Future future : futures) sum += (Long) future.get();
 
     System.out.println("Speed : " + (sum / futures.size()) + " ns per document.");
   }
 
   @AfterClass
   public void deinit() {
-    if (database != null)
-      database.drop();
+    if (database != null) database.drop();
   }
 
   private final class Saver implements Callable<Long> {
 
-    private Saver() {
-    }
+    private Saver() {}
 
     @Override
     public Long call() throws Exception {
@@ -119,7 +120,8 @@ public class LocalMTCreateDocumentSpeedTest {
 
         final String user = users.get(random.nextInt(users.size()));
 
-        final OPartitionedDatabasePool pool = poolFactory.get(System.getProperty("url"), user, user);
+        final OPartitionedDatabasePool pool =
+            poolFactory.get(System.getProperty("url"), user, user);
         final ODatabaseDocumentTx database = pool.acquire();
 
         ODocument record = new ODocument("Account");
@@ -142,8 +144,8 @@ public class LocalMTCreateDocumentSpeedTest {
 
   private final class Reader implements Callable<Long> {
 
-    private final int   docCount;
-    private final int   clusterId;
+    private final int docCount;
+    private final int clusterId;
     public volatile int size;
 
     public Reader(long docCount, int clusterId) {
@@ -155,15 +157,15 @@ public class LocalMTCreateDocumentSpeedTest {
     public Long call() throws Exception {
 
       latch.await();
-      final OPartitionedDatabasePool pool = poolFactory.get(System.getProperty("url"), "admin", "admin");
+      final OPartitionedDatabasePool pool =
+          poolFactory.get(System.getProperty("url"), "admin", "admin");
       final ODatabaseDocumentTx database = pool.acquire();
 
       long counter = 0;
       long start = System.nanoTime();
       while (!stop) {
         ODocument document = database.load(new ORecordId(clusterId, random.nextInt(docCount)));
-        if (document != null)
-          counter++;
+        if (document != null) counter++;
       }
       long end = System.nanoTime();
 
@@ -171,5 +173,4 @@ public class LocalMTCreateDocumentSpeedTest {
       return ((end - start) / counter);
     }
   }
-
 }
